Update cal.com tests

+++ library/querying/src/main/java/cz/matfyz/querying/core/QueryExecution.java
@@ -5,7 +5,7 @@ import cz.matfyz.core.querying.ListResult;
 public record QueryExecution(
     ListResult result,
     /** Time of parsing and planning. */
-    double planningTimeInMs,
+    long planningTimeInMs,
     /** Time of evaluating the query plan in our system (selection and projection). */
-    double evaluationTimeInMs
+    long evaluationTimeInMs
 ) {}
